feature: add ex29.py to demonstrate dictionary access.
This solution demonstrates accessing a dictionary with keys
that may not exist. The `get` method of the dictionary class
will return a value of `None` if the key does not exist. This is
a safe approach for accessing a dictionary because it does not
throw an exception if the key doesn't exist.

This completes exercise 29 from section 14.